refactor: pages/components 에서 pages/ 로 경로를 이동하여 componenets 폴더 제거
