+++ frontend/src/pages/indexPage/indexPage.styled.ts
