+++ frontend/src/pages/notFoundPage/notFoundPage.styled.ts
