+++ frontend/src/pages/resultPage/ResultPage.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router';
 
-import FallBackPage from '@pages/components/fallBackPage/FallBackPage';
+import FallBackPage from '@pages/fallBackPage/FallBackPage';
 
 import BaseLoading from '@features/loading/components/baseLoading/BaseLoading';
 import Header from '@features/map/components/header/Header';
