+++ frontend/src/pages/resultPage/resultPage.styled.ts
